+++ src/App.tsx
@@ -1,33 +1,21 @@
-import {
-  Avatar,
-  createTheme,
-  Grid,
-  NextUIProvider,
-  Spacer,
-} from "@nextui-org/react";
-import { useState } from "react";
+import { Avatar, Card, Grid, NextUIProvider, Spacer } from "@nextui-org/react";
+import { useMemo } from "react";
 import { Github, Moon, Sun } from "./assets";
-import "./App.less";
 import SearchInput from "./components/SearchInput";
-
-const lightTheme = createTheme({ type: "light" });
-const darkTheme = createTheme({ type: "dark" });
+import BackgroundSwitcher from "./components/BackgroundSwitcher";
+import { BgItem, formatItemToBackground } from "./utils";
+import { useBg } from "./hooks/useBg";
+import { useTheme } from "./hooks/useTheme";
+import "./App.less";
 
 function App() {
-  const [theme, setTheme] = useState(
-    window.localStorage.getItem("theme") || "dark"
-  );
-  const isDark = theme === "dark";
+  const [{ themeStr, theme }, { toggleColorScheme }] = useTheme();
+  const [{ bg }, { updateBgId, updateBgList }] = useBg();
 
-  const toggleColorScheme = () => {
-    const nextTheme = isDark ? "light" : "dark";
-
-    window.localStorage.setItem("theme", nextTheme);
-    setTheme(nextTheme);
-  };
+  const bgItem = useMemo(() => bg.list.find((item) => item.id === bg.id), [bg]);
 
   return (
-    <NextUIProvider theme={isDark ? darkTheme : lightTheme}>
+    <NextUIProvider theme={theme}>
       <Grid.Container
         className="app"
         direction="column"
@@ -36,6 +24,7 @@ function App() {
           h: "100vh",
           p: 32,
           ov: "auto",
+          bg: bg.id ? formatItemToBackground(bgItem as BgItem) : "",
         }}
       >
         <Grid.Container alignItems="center" justify="space-between">
@@ -43,16 +32,18 @@ function App() {
             <Avatar src="../logo.jpg" size="xl" />
           </Grid>
           <Grid className="operation">
-            <a
-              href="https://github.com/learnsomesome/curry_homepage"
-              target="blank"
-              title="Github"
-            >
-              <Github />
-            </a>
-            <a title="Toggle Color Scheme" onClick={toggleColorScheme}>
-              {isDark ? <Sun /> : <Moon />}
-            </a>
+            <Card css={{ fd: "row", p: 12 }}>
+              <a
+                href="https://github.com/learnsomesome/curry_homepage"
+                target="blank"
+                title="Github"
+              >
+                <Github />
+              </a>
+              <a title="Toggle Color Scheme" onClick={toggleColorScheme}>
+                {themeStr === "dark" ? <Sun /> : <Moon />}
+              </a>
+            </Card>
           </Grid>
         </Grid.Container>
         <Spacer y={12} />
@@ -64,6 +55,11 @@ function App() {
           <SearchInput />
         </Grid.Container>
       </Grid.Container>
+      <BackgroundSwitcher
+        bg={bg}
+        updateBgId={updateBgId}
+        updateBgList={updateBgList}
+      />
     </NextUIProvider>
   );
 }